pyproject.toml
[project]
name = "quan-ly-du-an"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "gspread>=6.2.1",
    "python-dotenv>=1.2.1",
]
